Progress 2

+++ src/app/admin/admin.component.ts
@@ -41,9 +41,11 @@ export class AdminComponent {
 	}
 
 	deleteAdvertById(advertId: string) {
-		this.advertService.deleteAdvert(advertId).pipe(untilDestroyed(this)).subscribe(() => {
-			window.alert('Inzerát bol úspešne vymazaný.');
-		}) 
+		if (window.confirm(`Naozaj chcete vymazať tento inzerát?\n${advertId}`)) {
+			this.advertService.deleteAdvert(advertId).pipe(untilDestroyed(this)).subscribe(() => {
+				window.alert('Inzerát bol úspešne vymazaný.');
+			});
+		}
 	}
 
 	copy(text: any): void {
